+++ plugins/woocommerce-admin/client/customize-store/assembler-hub/sidebar/global-styles/index.ts
@@ -1,2 +1,2 @@
-export { ColorPalette } from './color-palette-variations/color-palette';
+export { ColorPalette } from './color-palette-variations';
 export { ColorPanel } from './color-panel';
